Parsing fails when a number value is greater than Java Integer.MAX
Fixes #341

+++ yagi/src/main/java/org/raml/yagi/framework/grammar/rule/MaximumValueRule.java
@@ -61,7 +61,7 @@ public class MaximumValueRule extends Rule
         {
             value = ((FloatingNode) node).getValue();
         }
-        return value != null && value.compareTo(new BigDecimal(maximumValue.floatValue())) <= 0;
+        return value != null && value.compareTo(new BigDecimal(maximumValue.doubleValue())) <= 0;
     }
 
     @Override
